Callback once task completed
Added the ability to specify a callback in the grunt options. This will
then get run once the task has completed.

I have updated all the references to allDone() to instead call the
wrapper function completeTask which will execute the callback should one
be supplied. I have also updated the readme to include this new option
with an example.

The reason for this change set was to improve continuous integration, as
currently I have a use case that will generate a web font and use the
md5 hash in the filename to help prevent caching. This is then included
in a sass file as a font-face declaration. Now whenever the font is
updated, the new filename needs to be copied and pasted into the sass
file and the sass compiler needs to run. This allows me to write the
file name to a json file and then use the ruby gem "SASS-Vars" to import
json files which then creates a SASS variable which i can then use in my
font-declaration.